Update game system Now only one route is used : /game
